Cleanup outdated comments, rename directory, prebuild steps
