+++ packages/api/package.json
@@ -8,7 +8,8 @@
     "predev": "pnpm run init",
     "dev": "vite-node dev-server.mts",
     "test": "vitest",
-    "build": "tsc && rm -rf ./dist/drizzle && cp -R ./drizzle ./dist/drizzle && cp -R ./srcbooks/examples ./dist/srcbooks/examples",
+    "prebuild": "rm -rf ./dist",
+    "build": "tsc && cp -R ./drizzle ./dist/drizzle && cp -R ./srcbook/examples ./dist/srcbook/examples",
     "check-types": "tsc",
     "depcheck": "depcheck",
     "generate": "drizzle-kit generate",
