+++ packages/api/srcbook/examples/getting-started.srcmd
