+++ packages/api/srcbook/examples/websockets.srcmd
